Restructured the repository. Added setup.py and requirements.txt. Updated README.md
