+++ .gitignore
@@ -4,3 +4,4 @@ results*
 logs
 *__pycache__
 *.root
+*.aido*
