+++ example.py
@@ -1,11 +1,11 @@
 import os
-from SimulationWrapper import AIDO
-from simulation.SimulationHelpers import SimulationParameter, SimulationParameterDictionary
-from interface import AIDOUserInterfaceExample
+from modules.aido import AIDO  # required
+from modules.simulation_helpers import SimulationParameter, SimulationParameterDictionary  # required
+from container_examples.calo_opt.calo_opt_interface import AIDOUserInterfaceExample  # Import your derived class
 
 
 if __name__ == "__main__":
-    os.system("rm ./results -rf")
+    os.system("rm ./results -rf")  # remove everything from results
 
     sim_param_dict = SimulationParameterDictionary([
         SimulationParameter('thickness_absorber_0', 1.0, min_value=1E-3, max_value=5.0, sigma=0.2),
