+++ requirements.txt
@@ -1,0 +1,30 @@
+pip>=24.0
+b2luigi==1.0.1
+cachetools==5.4.0
+colorama==0.4.6
+cramjam==2.8.3
+decorator==5.1.1
+docutils==0.21.2
+fastparquet==2024.5.0
+fsspec==2024.6.1
+gitdb==4.0.11
+GitPython==3.1.43
+Jinja2==3.1.4
+lockfile==0.12.2
+luigi==3.5.1
+MarkupSafe==2.1.5
+numpy==2.0.1
+packaging==24.1
+pandas==2.2.2
+parse==1.20.2
+pyarrow==17.0.0
+python-daemon==3.0.1
+python-dateutil==2.9.0.post0
+pytz==2024.1
+retry2==0.9.5
+setuptools==71.1.0
+six==1.16.0
+smmap==5.0.1
+tenacity==8.5.0
+tornado==6.4.1
+tzdata==2024.1
